# apps/desktop/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hashlark API",
    "description": "Search many torrent indexers at once. All routes except /health and /openapi.json need `Authorization: Bearer <token>`.",
    "contact": {
      "name": "Hashlark contributors"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Lists every provider with its state and health.",
        "operationId": "list_providers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Adds a provider. Re-adding a definition with the same id updates it.",
        "operationId": "add_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProvider"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid definition; `details` lists the problems",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/resolve": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Gets the magnet link or `.torrent` URL for a result.",
        "operationId": "resolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTarget"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The provider failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddProvider": {
        "oneOf": [
          {
            "type": "object",
            "description": "From a YAML definition.",
            "required": [
              "yaml",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "definition"
                ]
              },
              "yaml": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "A Torznab endpoint (Jackett, Prowlarr, Bitmagnet).",
            "required": [
              "name",
              "url",
              "kind"
            ],
            "properties": {
              "api_key": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "kind": {
                "type": "string",
                "enum": [
                  "torznab"
                ]
              },
              "name": {
                "type": "string"
              },
              "url": {
                "type": "string"
              }
            }
          }
        ],
        "description": "How to add a provider."
      },
...
      "Capabilities": {
        "type": "object",
        "description": "What kinds of query a provider supports.",
        "required": [
          "text_search",
          "imdb_search",
          "paging"
        ],
        "properties": {
          "imdb_search": {
            "type": "boolean"
          },
          "paging": {
            "type": "boolean"
          },
          "text_search": {
            "type": "boolean"
          }
        }
      },
      "Category": {
        "type": "string",
        "description": "Content category. Providers map their own category ids onto these.",
        "enum": [
          "movies",
          "tv",
          "music",
          "books",
          "software",
          "games",
          "anime",
          "other"
        ]
      },
...
      "DownloadTarget": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "url",
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "magnet"
                ]
              },
              "url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "url",
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "torrent_file"
                ]
              },
              "url": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        ],
        "description": "What to hand to the OS to start a download."
      },
      "ErrorBody": {
        "type": "object",
        "description": "Body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "For `invalid_definition`: every problem found."
          },
          "message": {
            "type": "string"
          },
          "provider_error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorKind",
                "description": "For `provider_error`: what went wrong with the provider."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ErrorKind": {
        "type": "string",
        "description": "Stable, serializable category of a [`ProviderError`].",
        "enum": [
          "timeout",
          "blocked",
          "challenge_required",
          "auth_failed",
          "rate_limited",
          "parse_failed",
          "http",
          "network"
        ]
      },
...
      "HealthState": {
        "type": "string",
        "description": "Overall state shown next to a provider.",
        "enum": [
          "unknown",
          "ok",
          "degraded",
          "failing",
          "auto_disabled"
        ]
      },
      "HealthSummary": {
        "type": "object",
        "required": [
          "state",
          "consecutive_failures"
        ],
        "properties": {
          "consecutive_failures": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "disabled_until": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Unix ms until which the provider is skipped, if auto-disabled."
          },
          "last_checked_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Unix ms of the last attempt."
          },
          "last_error_kind": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "p50_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "p95_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "state": {
            "$ref": "#/components/schemas/HealthState"
          },
          "success_rate": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Share of successful attempts in the recent window."
          }
        }
      },
...
      "NetworkPolicy": {
        "type": "object",
        "description": "Per-provider network policy.",
        "properties": {
          "proxy": {
            "type": [
              "string",
              "null"
            ],
            "description": "Used when `route` is `proxy`.",
            "default": null
          },
          "route": {
            "$ref": "#/components/schemas/Route",
            "default": "default"
          }
        }
      },
...
      "ProviderKind": {
        "type": "string",
        "description": "How a provider is implemented (see ADR 0005).",
        "enum": [
          "native",
          "definition",
          "torznab"
        ]
      },
...
      "ProviderSource": {
        "oneOf": [
          {
            "type": "object",
            "description": "Compiled into Hashlark.",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "native"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "version",
              "links",
              "type"
            ],
            "properties": {
              "links": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "repo_id": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Repository it came from; `null` if imported by hand."
              },
              "site_type": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SiteType"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": {
                "type": "string",
                "enum": [
                  "definition"
                ]
              },
              "version": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "url",
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "torznab"
                ]
              },
              "url": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Where a provider comes from."
      },
...
      "ProviderView": {
        "type": "object",
        "description": "A provider as shown to users.",
        "required": [
          "id",
          "name",
          "kind",
          "description",
          "enabled",
          "builtin",
          "categories",
          "health",
          "network_policy",
          "source",
          "settings"
        ],
        "properties": {
          "builtin": {
            "type": "boolean",
            "description": "Shipped with Hashlark: can be disabled but not removed."
          },
          "capabilities": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Capabilities"
              },
              {
                "type": "null"
              }
            ]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the provider couldn't be loaded (e.g. an invalid definition)."
          },
          "health": {
            "$ref": "#/components/schemas/HealthSummary"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "name": {
            "type": "string"
          },
          "network_policy": {
            "$ref": "#/components/schemas/NetworkPolicy"
          },
          "settings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingView"
            }
          },
          "source": {
            "$ref": "#/components/schemas/ProviderSource"
          }
        }
      },
...
      "ResolveRequest": {
        "type": "object",
        "required": [
          "result_id"
        ],
        "properties": {
          "prefer": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TargetKind",
                "description": "Ask for a magnet or a `.torrent` when the result allows it."
              },
              {
                "type": "null"
              }
            ]
          },
          "result_id": {
            "type": "string"
          }
        }
      },
      "Route": {
        "type": "string",
        "description": "How a provider's traffic is routed.",
        "enum": [
          "default",
          "direct",
          "proxy",
          "tor"
        ]
      },
...
      "SettingKind": {
        "type": "string",
        "enum": [
          "text",
          "password",
          "checkbox",
          "select"
        ]
      },
      "SettingView": {
        "type": "object",
        "description": "A user-configurable setting of a provider.",
        "required": [
          "name",
          "label",
          "kind",
          "required",
          "options",
          "is_set"
        ],
        "properties": {
          "is_set": {
            "type": "boolean",
            "description": "Whether a value is stored (useful for passwords)."
          },
          "kind": {
            "$ref": "#/components/schemas/SettingKind"
          },
          "label": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "required": {
            "type": "boolean"
          },
          "value": {
            "type": [
              "string",
              "null"
            ],
            "description": "Current value; always `null` for passwords."
          }
        }
      },
...
      "SiteType": {
        "type": "string",
        "enum": [
          "public",
          "semi_private",
          "private"
        ]
      },
...
      "TargetKind": {
        "type": "string",
        "description": "Which kind of [`DownloadTarget`] a caller would rather have.",
        "enum": [
          "magnet",
          "torrent_file"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}